+++ lib/src/builders/code_block_builder.dart
@@ -1,19 +1,20 @@
 import 'package:flutter/material.dart';
 
 import '../definition.dart';
-import '../helpers.dart';
+import '../helpers/utils.dart';
+import '../widgets/copy_button.dart';
 import 'builder.dart';
 
 class CodeBlockBuilder extends MarkdownElementBuilder {
   CodeBlockBuilder({
     TextStyle? textStyle,
-    this.codeblockPadding,
-    this.codeblockDecoration,
+    this.padding,
+    this.decoration,
     this.highlightBuilder,
   }) : super(textStyle: textStyle);
 
-  final EdgeInsets? codeblockPadding;
-  final BoxDecoration? codeblockDecoration;
+  final EdgeInsets? padding;
+  final BoxDecoration? decoration;
   final MarkdownHighlightBuilder? highlightBuilder;
 
   @override
@@ -28,11 +29,12 @@ class CodeBlockBuilder extends MarkdownElementBuilder {
   @override
   TextSpan buildText(text, parent, selectable) {
     final textContent = text.trimRight();
+    final style = const TextStyle(fontFamily: 'monospace').merge(parent.style);
 
     if (highlightBuilder == null) {
       return TextSpan(
         text: textContent,
-        style: parent.style,
+        style: style,
         mouseCursor: mouseCursor(selectable),
       );
     }
@@ -49,27 +51,42 @@ class CodeBlockBuilder extends MarkdownElementBuilder {
 
     return TextSpan(
       children: spans,
-      style: parent.style,
+      style: style,
       mouseCursor: mouseCursor(selectable),
     );
   }
 
   @override
   Widget buildWidget(element) {
-    final child = element.children.isNotEmpty
-        ? element.children.single
-        : const SizedBox.shrink();
+    const defaultPadding = EdgeInsets.all(8.0);
+    final defaultDecoration = BoxDecoration(
+      color: const Color(0xfff0f0f0),
+      borderRadius: BorderRadius.circular(5),
+    );
+
+    Widget child;
+    if (element.children.isNotEmpty) {
+      final textWidget = element.children.single;
+      child = Stack(
+        children: [
+          Scrollbar(
+            child: SingleChildScrollView(
+              scrollDirection: Axis.horizontal,
+              padding: padding ?? defaultPadding,
+              child: textWidget,
+            ),
+          ),
+          Positioned(right: 0, top: 0, child: CopyButton(textWidget)),
+        ],
+      );
+    } else {
+      child = const SizedBox(height: 15);
+    }
 
     return Container(
       width: double.infinity,
-      decoration: codeblockDecoration,
-      child: Scrollbar(
-        child: SingleChildScrollView(
-          scrollDirection: Axis.horizontal,
-          padding: codeblockPadding,
-          child: child,
-        ),
-      ),
+      decoration: decoration ?? defaultDecoration,
+      child: child,
     );
   }
 }