+++ lib/src/helpers/utils.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/rendering.dart';
+
+MouseCursor? mouseCursor(bool selectable) =>
+    selectable ? SystemMouseCursors.text : null;